+++ src/MiniWeb.AssetStorage.FileSystem/MiniWeb.AssetStorage.FileSystem.csproj
@@ -11,5 +11,4 @@
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
 
-
 </Project>
