+++ src/MiniWeb.AssetStorage.FileSystem/MiniWebAssetFilesystemStorage.cs
@@ -27,7 +27,6 @@ namespace MiniWeb.AssetStorage.FileSystem
 			Logger = logger;
 		}
 
-
 		public Task<IEnumerable<IAsset>> GetAllAssets()
 		{
 			string folder = Path.Combine(HostingEnvironment.WebRootPath, Configuration.AssetRootPath);
@@ -89,6 +88,10 @@ namespace MiniWeb.AssetStorage.FileSystem
 		private string[] ImageExtensions = new[] { ".png", ".jpg", ".jpeg", ".gif", ".bmp" };
 		public IWebHostEnvironment HostingEnvironment { get; }
 		public MiniWebAssetFileSystemConfig Configuration { get; }
+		public string VirtualPath { get; set; }
+		public string FileName { get; set; }
+		public string Folder { get; set; }
+		
 		public FileSystemAsset(IWebHostEnvironment env, MiniWebAssetFileSystemConfig config, string fullPath)
 		{
 			HostingEnvironment = env;
@@ -97,6 +100,7 @@ namespace MiniWeb.AssetStorage.FileSystem
 			FileName = Path.GetFileName(VirtualPath);
 			Folder = Path.GetDirectoryName(VirtualPath).Replace("\\", "/");
 		}
+
 		public FileInfo Info
 		{
 			get
@@ -106,7 +110,6 @@ namespace MiniWeb.AssetStorage.FileSystem
 			}
 		}
 
-		public string VirtualPath { get; set; }
 		public AssetType Type
 		{
 			get
@@ -119,9 +122,6 @@ namespace MiniWeb.AssetStorage.FileSystem
 				return AssetType.File;
 			}
 		}
-
-		public string FileName { get; set; }
-		public string Folder { get; set; }
 
 		private string GetVirtualPath(string path)
 		{
