+++ src/MiniWeb.Core/MiniWeb.Core.csproj
@@ -15,6 +15,5 @@
     <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="6.0.0" />
     <PackageReference Include="System.Runtime.Serialization.Primitives" Version="4.3.0" />
-
   </ItemGroup>
 </Project>
