+++ src/MiniWeb.StaticSiteGenerator/ConsoleArgumentAttribute.cs
@@ -1,0 +1,12 @@
+namespace MiniWeb.StaticSiteGenerator
+{
+    internal class ConsoleArgumentAttribute : Attribute
+    {
+        public ConsoleArgumentAttribute(string prefix)
+        {
+            CommandPrefix = prefix;
+        }
+
+        public string CommandPrefix { get; }
+    }
+}
